fix: работа над ошибками
