+++ src/shared/libs/utils.ts
@@ -1,6 +1,7 @@
 import { FRONT_URL } from 'config/api-config';
 import differenceInMilliseconds from 'date-fns/differenceInMilliseconds';
 import { Task } from 'entities/task/types';
+import { IFilterValues } from 'features/filter/types';
 
 export const isTaskUrgent = (date: string): boolean =>
   differenceInMilliseconds(new Date(date), new Date()) < 86400000;
@@ -50,4 +51,56 @@ export const sortTasks = (
     return 0;
   });
   return sortedTasks;
+};
+
+export const handleFilterTasks = (
+  tasks: Task[],
+  setFilterTasks: (date: Task[]) => void,
+  infoFilterTasks: IFilterValues
+) => {
+  const handleTasksFilter = (arr: Task[]) =>
+    arr.filter((task: Task) =>
+      infoFilterTasks.categories.includes(task.category.name)
+    );
+
+  if (tasks) {
+    setFilterTasks(tasks);
+  }
+  const sortDisplay = (arr: Task[], text: string): Task[] => {
+    let sortedTasks: Task[] = [];
+    switch (text) {
+      case 'date':
+        sortedTasks = sortTasks(arr, 'date');
+        break;
+      case 'decreasingPoints':
+        sortedTasks = sortTasks(arr, 'decreasing');
+        break;
+      case 'increasingPoints':
+        sortedTasks = sortTasks(arr, 'increasing');
+        break;
+      default:
+        sortedTasks = arr;
+        break;
+    }
+    return sortedTasks;
+  };
+  if (infoFilterTasks?.categories.length) {
+    const filteredTasks = tasks.filter((task: Task) => {
+      return infoFilterTasks.categories.includes(task.category.name);
+    });
+    if (infoFilterTasks?.sortBy) {
+      sortDisplay(handleTasksFilter(tasks), infoFilterTasks.sortBy);
+    } else {
+      setFilterTasks(filteredTasks);
+    }
+  }
+  if (infoFilterTasks?.sortBy) {
+    if (infoFilterTasks?.categories.length > 0) {
+      setFilterTasks(
+        sortDisplay(handleTasksFilter(tasks), infoFilterTasks.sortBy)
+      );
+    } else {
+      setFilterTasks(sortDisplay(tasks, infoFilterTasks.sortBy));
+    }
+  }
 };
